Rename ajax.ajaxError to ajax.ajax_error
Camelcase format is not compatible with HTML data-* attributes.

I keep error translations inside HTML data-* attributes, like so:

```html
<div class="js-form-translations"
    data-form-required="This field is required"
    data-form-useremail-ajax="Checking.."
    data-form-useremail-ajax_error="Email already exists"
></div>
```

and dynamically parse and extend IdealForms errors object with the values
found in `.js-form-translations`.

Btw the approach of keeping errors in HTML is very convenient, especially
in app that is designed so that translators' UI displays all terms used in
currently viewed View ready to be translated. If translations are kept in JS,
that would be harder to accomplish.

Without this patch, the workaround is dirty:

```js
...

if( key === 'ajax_error' ) key = 'ajaxError';

...
```